docs: laugh a little less
The paragraph is certainly still not written in the best way, but I have not come up with anything better yet. I think it is necessary as an addition to the second one (people tend to have low trust in those who appear out of nowhere, especially nowadays, when GPT-like tools are widely available). The question is, of course, a good one: why am I justifying myself to anyone at all?